+++ packages/synapse-interface/components/_Transaction/components/TimeRemaining.tsx
@@ -1,26 +1,38 @@
 import { useMemo } from 'react'
+import ExclamationIcon from '@components/icons/ExclamationIcon'
 
 /**
- * @param isComplete: Transaction has been marked complete
- * @param remainingTime: Time remaining for Transaction, in seconds
  * @param isDelayed: Elapsed Time is over Estimated Time for Transaction
+ * @param remainingTime: Time remaining for Transaction, in seconds
  * @param delayedTime: Delayed Time, in seconds
+ * @param status: Transaction status
  * @returns Remaining time in minutes (in seconds if < 1 min) if not delayed.
  * Otherwise, return text representative of delayed transaction.
  */
 export const TimeRemaining = ({
-  isComplete,
-  remainingTime,
   isDelayed,
+  remainingTime,
   delayedTime,
+  status,
 }: {
-  isComplete: boolean
-  remainingTime: number
   isDelayed: boolean
+  remainingTime: number
   delayedTime: number | null
+  status: string
 }) => {
+  const isComplete = status === 'completed'
+  const isReverted = status === 'reverted'
+
   if (isComplete) {
     return <div className="text-sm text-green-400">Complete!</div>
+  }
+
+  if (isReverted) {
+    return (
+      <span className="flex items-center space-x-1 text-sm">
+        <ExclamationIcon className="w-4 h-4" /> <span>Reverted</span>
+      </span>
+    )
   }
 
   if (isDelayed) {